+++ R/autogif.R
@@ -5,25 +5,38 @@
 #' @importFrom vroom vroom
 NULL
 
-#' Make a gif from a TROLL simulation
+#' Create an animation from a TROLL simulation
+#'
+#' `autogif()` uses `ggplot2` and `gganimate` to render an animation from a
+#' TROLL simulation. The animation can include a vertical cut in the forest
+#' structure along the X-axis highlighting either tree species (`variables =
+#' 'species'`), either tree height relative to their maximum height (`variables
+#' = 'height_ct'`), or tree carbon acquisition with net over growth primary
+#' productivity (`variables = 'npp_gpp'`). The animation can also include a top
+#' view of the canopy representing either canopy trees height (`variables =
+#' 'height'`) or total leaf area index per pixel (`variables = 'lai'`).
 #'
 #' @param name char. Model name (if NULL timestamp).
 #' @param path char. Path to save the simulation outputs, the default is null
-#'   corresponding to a simulation in memory without saved intermediary files.
+#'   corresponding to a simulation in memory without saved intermediary files
+#'   (based on temporary files from [option.rcontroll]).
 #' @param variables char. Variables to build as a gif among 'species',
-#'   'height_ct', 'npp_gpp', 'height', or 'lai' (see details).
-#' @param global df. Global parameters.
-#' @param species df. Species parameters.
-#' @param climate df. Climate parameters.
-#' @param daily df. Daily variation parameters.
+#'   'height_ct', 'npp_gpp', 'height', or 'lai'.
+#' @param global df. Global parameters (e.g. [TROLLv3_input] or using
+#'   [generate_parameters()]).
+#' @param species df. Species parameters (e.g. [TROLLv3_species]).
+#' @param climate df. Climate parameters (e.g. [TROLLv3_climatedaytime12]).
+#' @param daily df. Daily variation parameters (e.g. [TROLLv3_daytimevar]).
 #' @param forest df. TROLL with forest input, if null starts from an empty grid
-#'   (default NULL).
+#'   (default NULL) (e.g. using [TROLLv3_output] with [get_forest()]).
 #' @param verbose bool. Show TROLL outputs in the console.
 #' @param overwrite bool. Overwrite previous outputs.
 #' @param thin int. Vector of integers corresponding to the iterations to be
 #'   kept to reduce output size, default is NULL and corresponds to no thinning.
 #'
-#' @return A list of gganimate objects corresponding to chosen outputs.
+#' @return A list of `gganimate` objects corresponding to chosen outputs.
+#'
+#' @seealso [autoplot,trollsim-method]
 #'
 #' @examples
 #' \dontrun{
@@ -46,8 +59,10 @@ NULL
 #' @export
 autogif <- function(name = NULL,
                     path = NULL,
-                    variables = c("species", "height_ct", "npp_gpp",
-                                  "height", "lai"),
+                    variables = c(
+                      "species", "height_ct", "npp_gpp",
+                      "height", "lai"
+                    ),
                     global,
                     species,
                     climate,
@@ -56,20 +71,24 @@ autogif <- function(name = NULL,
                     verbose = TRUE,
                     overwrite = TRUE,
                     thin = NULL) {
-  if (!all(variables %in% c("species", "height_ct",
-                            "npp_gpp", "height", "lai"))) {
-    stop("No valid autogif available for: ", variables,
-         ". Please use either 'species', 'height_ct', 
-         'npp_gpp', 'height', or 'lai'.")
+  if (!all(variables %in% c(
+    "species", "height_ct",
+    "npp_gpp", "height", "lai"
+  ))) {
+    stop(
+      "No valid autogif available for: ", variables,
+      ". Please use either 'species', 'height_ct',
+         'npp_gpp', 'height', or 'lai'."
+    )
   }
   if (global[which(global$param == "_OUTPUT_extended"), 2] == 0) {
-    stop("_OUTPUT_extended option should be activated in global 
-         parameters to produce gif from TROLL 
+    stop("_OUTPUT_extended option should be activated in global
+         parameters to produce gif from TROLL
          (generate_parameters(_OUTPUT_extended=1)).")
   }
   if (global[which(global$param == "extent_visual"), 2] == 0) {
-    stop("extent_visual option should be non null in global 
-         parameters to produce gif from TROLL 
+    stop("extent_visual option should be non null in global
+         parameters to produce gif from TROLL
          (e.g. generate_parameters(extent_visual=100)).")
   }
 
@@ -104,25 +123,37 @@ autogif <- function(name = NULL,
 }
 
 .troll_to_gif <- function(name,
-                       path,
-                       variables) {
-  LAI <- height <- height_spikefree <- iter <- NULL
-  ratio_NPP_GPP <- ratio_height_Ct <- sp_lab <- NULL
+                          path,
+                          variables) {
+  LAI <- height <- height_spikefree <- iter <- NULL # nolint
+  ratio_NPP_GPP <- ratio_height_Ct <- sp_lab <- NULL # nolint
   results <- list()
-  slice <- vroom(file.path(path, name,
-                           paste0(name, "_0_visual_slice.txt")),
-                 col_types = cols()) %>%
+  slice <- vroom(
+    file.path(
+      path, name,
+      paste0(name, "_0_visual_slice.txt")
+    ),
+    col_types = cols()
+  ) %>%
     mutate(sp_lab = as.factor(round(sp_lab))) %>%
     mutate(height = height + 0.5)
-  field <- vroom(file.path(path, name,
-                           paste0(name, "_0_visual_field.txt")),
-                 col_types = cols()) %>%
+  field <- vroom(
+    file.path(
+      path, name,
+      paste0(name, "_0_visual_field.txt")
+    ),
+    col_types = cols()
+  ) %>%
     mutate(col = col + 0.5, row = row + 0.5)
 
   if ("species" %in% variables) {
-    results$species <- ggplot(slice,
-                              aes(x = col, y = height,
-                                  fill = sp_lab, alpha = row)) +
+    results$species <- ggplot(
+      slice,
+      aes(
+        x = col, y = height,
+        fill = sp_lab, alpha = row
+      )
+    ) +
       scale_x_continuous(expand = c(0, 0)) +
       scale_y_continuous(expand = c(0, 0)) +
       scale_fill_viridis_d(option = "magma") +
@@ -140,9 +171,13 @@ autogif <- function(name = NULL,
 
 
   if ("npp_gpp" %in% variables) {
-    results$npp_gpp <- ggplot(data = slice,
-                              aes(x = col, y = height,
-                                  fill = ratio_NPP_GPP, alpha = row)) +
+    results$npp_gpp <- ggplot(
+      data = slice,
+      aes(
+        x = col, y = height,
+        fill = ratio_NPP_GPP, alpha = row
+      )
+    ) +
       scale_fill_viridis_c(expression(frac(NPP, GPP)), direction = -1) +
       scale_alpha(range = c(0.5, 1.0), guide = "none") +
       xlab("X (m)") +
@@ -158,10 +193,14 @@ autogif <- function(name = NULL,
 
 
   if ("height_ct" %in% variables) {
-    results$height_ct <- ggplot(data = slice,
-                                aes(x = col, y = height,
-                                    fill = ratio_height_Ct,
-                                    alpha = row)) +
+    results$height_ct <- ggplot(
+      data = slice,
+      aes(
+        x = col, y = height,
+        fill = ratio_height_Ct,
+        alpha = row
+      )
+    ) +
       scale_fill_viridis_c(expression(frac(height, Ct))) +
       scale_alpha(range = c(0.5, 1.0), guide = "none") +
       xlab("X (m)") +
@@ -177,9 +216,13 @@ autogif <- function(name = NULL,
 
 
   if ("height" %in% variables) {
-    results$height <- ggplot(data = field,
-                             aes(x = col, y = row,
-                                 fill = height_spikefree)) +
+    results$height <- ggplot(
+      data = field,
+      aes(
+        x = col, y = row,
+        fill = height_spikefree
+      )
+    ) +
       scale_fill_viridis_c("height (m)") +
       xlab("X (m)") +
       ylab("Y (m)") +